Annotation Refactor - Move annotations manager to src/annotations
