+++ src/content_script.js
@@ -3,7 +3,7 @@ import { RemoteFunctionRegistry } from './util/webextensionRPC'
 import 'src/activity-logger/content_script'
 import 'src/page-analysis/content_script'
 import 'src/search-injection/content_script'
-import AnnotationsManager from 'src/sidebar-overlay/annotations-manager'
+import AnnotationsManager from 'src/annotations/annotations-manager'
 import initContentTooltip from 'src/content-tooltip/content_script'
 import 'src/direct-linking/content_script'
 import initRibbonAndSidebar from './sidebar-overlay/content_script'
