+++ src/sidebar-overlay/content_script/ribbon-interactions.ts
@@ -10,7 +10,7 @@ import { getSidebarState } from '../utils'
 import { getTooltipState } from 'src/content-tooltip/utils'
 import { createRootElement, destroyRootElement } from './rendering'
 import { removeHighlights } from '../../highlighting/ui/highlight-interactions'
-import AnnotationsManager from 'src/sidebar-overlay/annotations-manager'
+import AnnotationsManager from 'src/annotations/annotations-manager'
 import ToolbarNotifications from 'src/toolbar-notification/content_script'
 import { insertTooltip, removeTooltip } from 'src/content-tooltip/interactions'
 import { RibbonInteractionsInterface } from 'src/sidebar-overlay/ribbon/types'
