+++ src/sidebar-overlay/sidebar/actions.ts
@@ -3,7 +3,7 @@ import { createAction } from 'redux-act'
 import { Thunk } from '../types'
 import { Page } from './types'
 import * as selectors from './selectors'
-import AnnotationsManager from '../annotations-manager'
+import AnnotationsManager from '../../annotations/annotations-manager'
 import { remoteFunction } from 'src/util/webextensionRPC'
 import { EVENT_NAMES } from 'src/analytics/internal/constants'
 import { FLOWS, STAGES } from 'src/overview/onboarding/constants'
