+++ src/social-integration/content_script.ts
@@ -1,4 +1,4 @@
-import AnnotationsManager from 'src/sidebar-overlay/annotations-manager'
+import AnnotationsManager from 'src/annotations/annotations-manager'
 import initTwitterIntegration from 'src/social-integration/twitter'
 
 export default function initSocialIntegration({
